// src/views/multi-cloud/elastic-ip/components/interface.ts
export interface IHooksBasic {
  regionList?: any[]
  chargeModeList?: any[]
  projectList?: any[]
  highConfig?: boolean
  lineList?: any[]
  bandwidthTypeOption?: any[]
  bandwidthList?: any[]
  sizeList?: any[]
  duration?: any[]
  exNetList?: any[]
  eipTypeDes?: string
  bandwidthDes?: string
  labelNum?: number
}